Porting Optical Flow, Segmentation, Video models (#5619)
* Porting raft

* Porting video resnet

* Porting deeplabv3

* Porting fcn and lraspp

* Fixing the tests and linter
