+++ torchvision/models/segmentation/_utils.py
@@ -4,7 +4,6 @@ from typing import Optional, Dict
 from torch import nn, Tensor
 from torch.nn import functional as F
 
-from ..._internally_replaced_utils import load_state_dict_from_url
 from ...utils import _log_api_usage_once
 
 
@@ -36,10 +35,3 @@ class _SimpleSegmentationModel(nn.Module):
             result["aux"] = x
 
         return result
-
-
-def _load_weights(arch: str, model: nn.Module, model_url: Optional[str], progress: bool) -> None:
-    if model_url is None:
-        raise ValueError(f"No checkpoint is available for {arch}")
-    state_dict = load_state_dict_from_url(model_url, progress=progress)
-    model.load_state_dict(state_dict)
