+++ torchvision/models/video/resnet.py
@@ -1,18 +1,25 @@
+from functools import partial
 from typing import Tuple, Optional, Callable, List, Sequence, Type, Any, Union
 
 import torch.nn as nn
 from torch import Tensor
 
-from ..._internally_replaced_utils import load_state_dict_from_url
+from ...transforms import VideoClassificationEval, InterpolationMode
 from ...utils import _log_api_usage_once
+from .._api import WeightsEnum, Weights
+from .._meta import _KINETICS400_CATEGORIES
+from .._utils import handle_legacy_interface, _ovewrite_named_param
 
-__all__ = ["r3d_18", "mc3_18", "r2plus1d_18"]
 
-model_urls = {
-    "r3d_18": "https://download.pytorch.org/models/r3d_18-b3b3357e.pth",
-    "mc3_18": "https://download.pytorch.org/models/mc3_18-a90a0ba3.pth",
-    "r2plus1d_18": "https://download.pytorch.org/models/r2plus1d_18-91a641e6.pth",
-}
+__all__ = [
+    "VideoResNet",
+    "R3D_18_Weights",
+    "MC3_18_Weights",
+    "R2Plus1D_18_Weights",
+    "r3d_18",
+    "mc3_18",
+    "r2plus1d_18",
+]
 
 
 class Conv3DSimple(nn.Conv3d):
@@ -281,80 +288,152 @@ class VideoResNet(nn.Module):
         return nn.Sequential(*layers)
 
 
-def _video_resnet(arch: str, pretrained: bool = False, progress: bool = True, **kwargs: Any) -> VideoResNet:
-    model = VideoResNet(**kwargs)
+def _video_resnet(
+    block: Type[Union[BasicBlock, Bottleneck]],
+    conv_makers: Sequence[Type[Union[Conv3DSimple, Conv3DNoTemporal, Conv2Plus1D]]],
+    layers: List[int],
+    stem: Callable[..., nn.Module],
+    weights: Optional[WeightsEnum],
+    progress: bool,
+    **kwargs: Any,
+) -> VideoResNet:
+    if weights is not None:
+        _ovewrite_named_param(kwargs, "num_classes", len(weights.meta["categories"]))
 
-    if pretrained:
-        state_dict = load_state_dict_from_url(model_urls[arch], progress=progress)
-        model.load_state_dict(state_dict)
+    model = VideoResNet(block, conv_makers, layers, stem, **kwargs)
+
+    if weights is not None:
+        model.load_state_dict(weights.get_state_dict(progress=progress))
+
     return model
 
 
-def r3d_18(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> VideoResNet:
+_COMMON_META = {
+    "task": "video_classification",
+    "publication_year": 2017,
+    "size": (112, 112),
+    "min_size": (1, 1),
+    "categories": _KINETICS400_CATEGORIES,
+    "interpolation": InterpolationMode.BILINEAR,
+    "recipe": "https://github.com/pytorch/vision/tree/main/references/video_classification",
+}
+
+
+class R3D_18_Weights(WeightsEnum):
+    KINETICS400_V1 = Weights(
+        url="https://download.pytorch.org/models/r3d_18-b3b3357e.pth",
+        transforms=partial(VideoClassificationEval, crop_size=(112, 112), resize_size=(128, 171)),
+        meta={
+            **_COMMON_META,
+            "architecture": "R3D",
+            "num_params": 33371472,
+            "acc@1": 52.75,
+            "acc@5": 75.45,
+        },
+    )
+    DEFAULT = KINETICS400_V1
+
+
+class MC3_18_Weights(WeightsEnum):
+    KINETICS400_V1 = Weights(
+        url="https://download.pytorch.org/models/mc3_18-a90a0ba3.pth",
+        transforms=partial(VideoClassificationEval, crop_size=(112, 112), resize_size=(128, 171)),
+        meta={
+            **_COMMON_META,
+            "architecture": "MC3",
+            "num_params": 11695440,
+            "acc@1": 53.90,
+            "acc@5": 76.29,
+        },
+    )
+    DEFAULT = KINETICS400_V1
+
+
+class R2Plus1D_18_Weights(WeightsEnum):
+    KINETICS400_V1 = Weights(
+        url="https://download.pytorch.org/models/r2plus1d_18-91a641e6.pth",
+        transforms=partial(VideoClassificationEval, crop_size=(112, 112), resize_size=(128, 171)),
+        meta={
+            **_COMMON_META,
+            "architecture": "R(2+1)D",
+            "num_params": 31505325,
+            "acc@1": 57.50,
+            "acc@5": 78.81,
+        },
+    )
+    DEFAULT = KINETICS400_V1
+
+
+@handle_legacy_interface(weights=("pretrained", R3D_18_Weights.KINETICS400_V1))
+def r3d_18(*, weights: Optional[R3D_18_Weights] = None, progress: bool = True, **kwargs: Any) -> VideoResNet:
     """Construct 18 layer Resnet3D model as in
     https://arxiv.org/abs/1711.11248
 
     Args:
-        pretrained (bool): If True, returns a model pre-trained on Kinetics-400
+        weights (R3D_18_Weights, optional): The pretrained weights for the model
         progress (bool): If True, displays a progress bar of the download to stderr
 
     Returns:
-        nn.Module: R3D-18 network
+        VideoResNet: R3D-18 network
     """
+    weights = R3D_18_Weights.verify(weights)
 
     return _video_resnet(
-        "r3d_18",
-        pretrained,
+        BasicBlock,
+        [Conv3DSimple] * 4,
+        [2, 2, 2, 2],
+        BasicStem,
+        weights,
         progress,
-        block=BasicBlock,
-        conv_makers=[Conv3DSimple] * 4,
-        layers=[2, 2, 2, 2],
-        stem=BasicStem,
         **kwargs,
     )
 
 
-def mc3_18(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> VideoResNet:
+@handle_legacy_interface(weights=("pretrained", MC3_18_Weights.KINETICS400_V1))
+def mc3_18(*, weights: Optional[MC3_18_Weights] = None, progress: bool = True, **kwargs: Any) -> VideoResNet:
     """Constructor for 18 layer Mixed Convolution network as in
     https://arxiv.org/abs/1711.11248
 
     Args:
-        pretrained (bool): If True, returns a model pre-trained on Kinetics-400
+        weights (MC3_18_Weights, optional): The pretrained weights for the model
         progress (bool): If True, displays a progress bar of the download to stderr
 
     Returns:
-        nn.Module: MC3 Network definition
+        VideoResNet: MC3 Network definition
     """
+    weights = MC3_18_Weights.verify(weights)
+
     return _video_resnet(
-        "mc3_18",
-        pretrained,
+        BasicBlock,
+        [Conv3DSimple] + [Conv3DNoTemporal] * 3,  # type: ignore[list-item]
+        [2, 2, 2, 2],
+        BasicStem,
+        weights,
         progress,
-        block=BasicBlock,
-        conv_makers=[Conv3DSimple] + [Conv3DNoTemporal] * 3,  # type: ignore[list-item]
-        layers=[2, 2, 2, 2],
-        stem=BasicStem,
         **kwargs,
     )
 
 
-def r2plus1d_18(pretrained: bool = False, progress: bool = True, **kwargs: Any) -> VideoResNet:
+@handle_legacy_interface(weights=("pretrained", R2Plus1D_18_Weights.KINETICS400_V1))
+def r2plus1d_18(*, weights: Optional[R2Plus1D_18_Weights] = None, progress: bool = True, **kwargs: Any) -> VideoResNet:
     """Constructor for the 18 layer deep R(2+1)D network as in
     https://arxiv.org/abs/1711.11248
 
     Args:
-        pretrained (bool): If True, returns a model pre-trained on Kinetics-400
+        weights (R2Plus1D_18_Weights, optional): The pretrained weights for the model
         progress (bool): If True, displays a progress bar of the download to stderr
 
     Returns:
-        nn.Module: R(2+1)D-18 network
+        VideoResNet: R(2+1)D-18 network
     """
+    weights = R2Plus1D_18_Weights.verify(weights)
+
     return _video_resnet(
-        "r2plus1d_18",
-        pretrained,
+        BasicBlock,
+        [Conv2Plus1D] * 4,
+        [2, 2, 2, 2],
+        R2Plus1dStem,
+        weights,
         progress,
-        block=BasicBlock,
-        conv_makers=[Conv2Plus1D] * 4,
-        layers=[2, 2, 2, 2],
-        stem=R2Plus1dStem,
         **kwargs,
     )
