+++ torchvision/prototype/__init__.py
@@ -1,5 +1,4 @@
 from . import datasets
 from . import features
-from . import models
 from . import transforms
 from . import utils
