--- torchvision/prototype/models/segmentation/__init__.py
@@ -1,3 +1,0 @@
-from .fcn import *
-from .lraspp import *
-from .deeplabv3 import *
